# docs/api/openapi.yaml
components:
  schemas:
...
    Body_upload_order_photos_api_v1_orders_photos_upload_post:
      properties:
        files:
          items:
            contentMediaType: application/octet-stream
            type: string
          title: Files
          type: array
      required:
      - files
      title: Body_upload_order_photos_api_v1_orders_photos_upload_post
      type: object
...
    ForbiddenCategory:
      enum:
      - intimate
      - weapons
      - drugs
      title: ForbiddenCategory
      type: string
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    OrderCreateRequest:
      properties:
        birth_date:
          anyOf:
          - type: string
          - type: 'null'
          title: Birth Date
        category:
          $ref: '#/components/schemas/ProductCategory'
        company_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Company Id
        customer_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Customer Name
        device_model:
          anyOf:
          - maxLength: 64
            type: string
          - type: 'null'
          title: Device Model
        forbidden_categories:
          default: []
          items:
            $ref: '#/components/schemas/ForbiddenCategory'
          title: Forbidden Categories
          type: array
        model_display_name:
          anyOf:
          - maxLength: 120
            type: string
          - type: 'null'
          title: Model Display Name
        os_version:
          anyOf:
          - maxLength: 64
            type: string
          - type: 'null'
          title: Os Version
        photos_prefix:
          anyOf:
          - type: string
          - type: 'null'
          description: Префикс ключей фото в MinIO (photos/{task_uuid}/)
          title: Photos Prefix
        promocode:
          anyOf:
          - type: string
          - type: 'null'
          title: Promocode
        receipt_email:
          anyOf:
          - type: string
          - type: 'null'
          title: Receipt Email
        scale_calibration:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Scale Calibration
        task_uuid:
          description: UUID для идемпотентности
          title: Task Uuid
          type: string
        tier:
          $ref: '#/components/schemas/Tier'
        upsell_options:
          default: []
          items:
            $ref: '#/components/schemas/UpsellOption'
          title: Upsell Options
          type: array
        zip_sha256:
          anyOf:
          - maxLength: 64
            minLength: 64
            type: string
          - type: 'null'
          title: Zip Sha256
      required:
      - category
      - tier
      - task_uuid
      title: OrderCreateRequest
      type: object
    OrderPayRequest:
      properties:
        customer_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Customer Name
        payment_method:
          default: redirect
          title: Payment Method
          type: string
      title: OrderPayRequest
      type: object
...
    PhotoEncryptionKeyBody:
      properties:
        algorithm:
          default: aes-256-gcm
          title: Algorithm
          type: string
        key_b64:
          title: Key B64
          type: string
        task_uuid:
          title: Task Uuid
          type: string
      required:
      - task_uuid
      - key_b64
      title: PhotoEncryptionKeyBody
      type: object
    PhotosPrepareRequest:
      properties:
        company_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Company Id
        task_uuid:
          anyOf:
          - type: string
          - type: 'null'
          title: Task Uuid
      title: PhotosPrepareRequest
      type: object
...
    ProductCategory:
      enum:
      - clothing
      - shoes
      - electronics
      - furniture
      - decor
      - toys
      - adult
      - other
      title: ProductCategory
      type: string
...
    Tier:
      enum:
      - small
      - large
      title: Tier
      type: string
...
    UpsellOption:
      enum:
      - real_scale
      - video_360
      - virtual_tryon
      - hole_filling
      title: UpsellOption
      type: string
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: API платформы 3D-моделей для маркетплейсов
  title: KWork Mob API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/orders:
    get:
      description: §3.5.3 / §3.16.2 — личные заказы или заказы компании с фильтром
        по сотруднику.
      operationId: list_orders_api_v1_orders_get
      parameters:
      - in: query
        name: company_id
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Company Id
      - description: Фильтр исполнитель §3.16.2
        in: query
        name: user_id
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: Фильтр исполнитель §3.16.2
          title: User Id
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: List Orders
      tags:
      - Заказы
  /api/v1/orders/create:
    post:
      operationId: create_order_api_v1_orders_create_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Create Order
      tags:
      - Заказы
  /api/v1/orders/photos/encryption-key:
    post:
      description: '§10.6.2: ключ E2E шифрования фото (отдельно от ciphertext в MinIO).'
      operationId: register_photo_encryption_key_api_v1_orders_photos_encryption_key_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhotoEncryptionKeyBody'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Register Photo Encryption Key
      tags:
      - Заказы
  /api/v1/orders/photos/prepare:
    post:
      description: Presigned PUT на photos/{task_uuid}/view_00…11.jpg + access_log
        (§10.7.7).
      operationId: prepare_photos_upload_api_v1_orders_photos_prepare_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhotosPrepareRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Prepare Photos Upload
      tags:
      - Заказы
  /api/v1/orders/photos/upload:
    post:
      description: 'Multipart: ровно 12 файлов → MinIO.'
      operationId: upload_order_photos_api_v1_orders_photos_upload_post
      parameters:
      - in: query
        name: task_uuid
        required: true
        schema:
          title: Task Uuid
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_order_photos_api_v1_orders_photos_upload_post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Upload Order Photos
      tags:
      - Заказы
...
  /api/v1/orders/{order_id}:
    get:
      operationId: get_order_api_v1_orders__order_id__get
      parameters:
      - in: path
        name: order_id
        required: true
        schema:
          title: Order Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get Order
      tags:
      - Заказы
  /api/v1/orders/{order_id}/cancel:
    post:
      operationId: cancel_order_api_v1_orders__order_id__cancel_post
      parameters:
      - in: path
        name: order_id
        required: true
        schema:
          title: Order Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Cancel Order
      tags:
      - Заказы
  /api/v1/orders/{order_id}/pay:
    post:
      operationId: pay_order_api_v1_orders__order_id__pay_post
      parameters:
      - in: path
        name: order_id
        required: true
        schema:
          title: Order Id
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/OrderPayRequest'
              - type: 'null'
              title: Body
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Pay Order
      tags:
      - Заказы
  /api/v1/orders/{order_id}/status:
    get:
      operationId: order_status_api_v1_orders__order_id__status_get
      parameters:
      - in: path
        name: order_id
        required: true
        schema:
          title: Order Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Order Status
      tags:
      - Заказы
...